requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.4.0

# LLM and Agent Framework
llama-index>=0.9.11
llama-index-llms-mistralai>=0.1.2

# OCR and API
mistralai>=0.0.7
requests>=2.31.0
pillow>=10.0.0

# UI
streamlit>=1.29.0

# Development & Testing
pytest>=7.4.0
black>=23.9.1
pylint>=2.17.5
pytest-cov>=4.1.0